Use normalized messages in latest grouping
The merge from issue-255 added the normalization code as a free-floating helper.
Move that behavior into the current v2 grouping mechanism instead, so
new projects and upgraded projects use normalized exception messages through
the normal grouping mechanism path.

See #440.
