+++ issues/grouping_mechanisms/after_2_4_0.py
@@ -1,6 +1,13 @@
+from django.utils.encoding import force_str
+
+from sentry.grouping.strategies.message import normalize_message_for_grouping
+
 from issues.utils import get_title_for_exception_type_and_value
 
 
-def default_issue_grouper(calculated_type, calculated_value, transaction):
+def default_issue_grouper(data, calculated_type, calculated_value, transaction):
+    if "exception" in data and data["exception"]:
+        calculated_value = normalize_message_for_grouping(force_str(calculated_value))
+
     title = get_title_for_exception_type_and_value(calculated_type, calculated_value)
     return title + " ⋄ " + transaction
