+++ issues/tests.py
@@ -43,7 +43,6 @@ from .factories import denormalized_issue_fields
 from .utils import (
     get_grouping_result_for_data,
     get_issue_grouper_for_data,
-    get_issue_grouper_for_data_with_normalized_exception_value,
     get_title_for_exception_type_and_value,
     get_type_and_value_for_data,
 )
@@ -1019,124 +1018,6 @@ class GroupingUtilsTestCase(DjangoTestCase):
             data["fingerprint"] = fingerprint
         return data
 
-    def test_empty_data(self):
-        self.assertEqual("Log Message: <no log message> ⋄ <no transaction>", get_issue_grouper_for_data({}))
-
-    def test_logentry_message_takes_precedence(self):
-        self.assertEqual("Log Message: msg: ? ⋄ <no transaction>", get_issue_grouper_for_data({"logentry": {
-            "message": "msg: ?",
-            "formatted": "msg: foobar",
-        }}))
-
-    def test_logentry_with_formatted_only(self):
-        self.assertEqual("Log Message: msg: foobar ⋄ <no transaction>", get_issue_grouper_for_data({"logentry": {
-            "formatted": "msg: foobar",
-        }}))
-
-    def test_logentry_with_transaction(self):
-        self.assertEqual("Log Message: msg ⋄ transaction", get_issue_grouper_for_data({
-            "logentry": {
-                "message": "msg",
-            },
-            "transaction": "transaction",
-        }))
-
-    def test_exception_empty_trace(self):
-        self.assertEqual("<unknown> ⋄ <no transaction>", get_issue_grouper_for_data({"exception": {
-            "values": [],
-        }}))
-
-    def test_exception_trace_no_data(self):
-        self.assertEqual("<unknown> ⋄ <no transaction>", get_issue_grouper_for_data({"exception": {
-            "values": [{}],
-        }}))
-
-    def test_exception_value_only(self):
-        self.assertEqual("Error: exception message ⋄ <no transaction>", get_issue_grouper_for_data({"exception": {
-            "values": [{"value": "exception message"}],
-        }}))
-
-    def test_exception_type_only(self):
-        self.assertEqual("KeyError ⋄ <no transaction>", get_issue_grouper_for_data({"exception": {
-            "values": [{"type": "KeyError"}],
-        }}))
-
-    def test_exception_type_value(self):
-        self.assertEqual("KeyError: exception message ⋄ <no transaction>", get_issue_grouper_for_data({"exception": {
-            "values": [{"type": "KeyError", "value": "exception message"}],
-        }}))
-
-    def test_exception_multiple_frames(self):
-        self.assertEqual("KeyError: exception message ⋄ <no transaction>", get_issue_grouper_for_data({"exception": {
-            "values": [{}, {}, {}, {"type": "KeyError", "value": "exception message"}],
-        }}))
-
-    def test_exception_transaction(self):
-        self.assertEqual("KeyError ⋄ transaction", get_issue_grouper_for_data({
-            "transaction": "transaction",
-            "exception": {
-                "values": [{"type": "KeyError"}],
-            }
-        }))
-
-    def test_exception_function_is_ignored_unless_specifically_synthetic(self):
-        # I make no value-judgement here on whether this is something we want to replicate in the future; as it stands
-        # this test just documents the somewhat surprising behavior that we inherited from GlitchTip/Sentry.
-        self.assertEqual("Error ⋄ <no transaction>", get_issue_grouper_for_data({
-            "exception": {
-                "values": [{
-                    "stacktrace": {
-                        "frames": [{"function": "foo"}],
-                    },
-                }],
-            },
-        }))
-
-    def test_synthetic_exception_only(self):
-        self.assertEqual("<unknown> ⋄ <no transaction>", get_issue_grouper_for_data({
-            "exception": {
-                "values": [{
-                    "mechanism": {"synthetic": True},
-                }],
-            },
-        }))
-
-    def test_synthetic_exception_ignores_value(self):
-        self.assertEqual("<unknown> ⋄ <no transaction>", get_issue_grouper_for_data({
-            "exception": {
-                "values": [{
-                    "mechanism": {"synthetic": True},
-                    "value": "the ignored value",
-                }],
-            },
-        }))
-
-    def test_exception_uses_function_when_top_level_exception_is_synthetic(self):
-        self.assertEqual("foo ⋄ <no transaction>", get_issue_grouper_for_data({
-            "exception": {
-                "values": [{
-                    "mechanism": {"synthetic": True},
-                    "stacktrace": {
-                        "frames": [{"function": "foo"}],
-                    },
-                }],
-            },
-        }))
-
-    def test_exception_with_non_string_value(self):
-        # In the GlitchTip code there is a mention of value sometimes containing a non-string value. Whether this
-        # happens in practice is unknown to me, but let's build something that can handle it.
-        self.assertEqual("KeyError: 123 ⋄ <no transaction>", get_issue_grouper_for_data({"exception": {
-            "values": [{"type": "KeyError", "value": 123}],
-        }}))
-
-    def test_simple_fingerprint(self):
-        self.assertEqual("fixed string", get_issue_grouper_for_data({"fingerprint": ["fixed string"]}))
-
-    def test_fingerprint_with_default(self):
-        self.assertEqual("Log Message: <no log message> ⋄ <no transaction> ⋄ fixed string",
-                         get_issue_grouper_for_data({"fingerprint": ["{{ default }}", "fixed string"]}))
-
     def test_grouping_result_for_explicit_fingerprint_is_mechanismless(self):
         result = get_grouping_result_for_data({"fingerprint": ["fixed string"]})
 
@@ -1188,8 +1069,8 @@ class GroupingUtilsTestCase(DjangoTestCase):
                 first_data = self._exception_event_data(exception_type, first_value)
                 second_data = self._exception_event_data(exception_type, second_value)
 
-                first_key = get_issue_grouper_for_data_with_normalized_exception_value(first_data)
-                second_key = get_issue_grouper_for_data_with_normalized_exception_value(second_data)
+                first_key = get_issue_grouper_for_data(first_data)
+                second_key = get_issue_grouper_for_data(second_data)
 
                 self.assertEqual(first_key, second_key)
                 self.assertIn(expected_placeholder, first_key)
@@ -1198,7 +1079,7 @@ class GroupingUtilsTestCase(DjangoTestCase):
         value = "Cannot publish <Consumer object at 0x7fbb00112233>"
         data = self._exception_event_data("SlowConsumerError", value)
 
-        grouping_key = get_issue_grouper_for_data_with_normalized_exception_value(data)
+        grouping_key = get_issue_grouper_for_data(data)
         calculated_type, calculated_value = get_type_and_value_for_data(data)
         title = get_title_for_exception_type_and_value(calculated_type, calculated_value)
 
@@ -1214,7 +1095,7 @@ class GroupingUtilsTestCase(DjangoTestCase):
 
         self.assertEqual(
             "fixed 0x7fbb00112233",
-            get_issue_grouper_for_data_with_normalized_exception_value(data),
+            get_issue_grouper_for_data(data),
         )
 
     def test_normalized_exception_grouping_changes_default_fingerprint_expansion(self):
@@ -1229,8 +1110,8 @@ class GroupingUtilsTestCase(DjangoTestCase):
             fingerprint=["{{ default }}", "fixed string"],
         )
 
-        first_key = get_issue_grouper_for_data_with_normalized_exception_value(first_data)
-        second_key = get_issue_grouper_for_data_with_normalized_exception_value(second_data)
+        first_key = get_issue_grouper_for_data(first_data)
+        second_key = get_issue_grouper_for_data(second_data)
 
         self.assertEqual(first_key, second_key)
         self.assertIn("SlowConsumerError: Cannot publish <Consumer object at <hex>>", first_key)
