+++ issues/utils.py
@@ -4,7 +4,6 @@ from django.utils.encoding import force_str
 
 from sentry.stacktraces.functions import get_function_name_for_frame
 from sentry.stacktraces.processing import get_crash_frame_from_event_data, get_crash_location
-from sentry.grouping.strategies.message import normalize_message_for_grouping
 from sentry.utils.safe import get_path, trim
 
 from sentry.utils.strings import strip
@@ -159,43 +158,18 @@ def get_grouping_result_for_data(data, calculated_type=None, calculated_value=No
         for part in fingerprint:
             if part == "{{ default }}":
                 used_default = True
-                parts.append(mechanism.grouper(calculated_type, calculated_value, transaction))
+                parts.append(mechanism.grouper(data, calculated_type, calculated_value, transaction))
             else:
                 parts.append(str(part))
 
         return GroupingResult(" ⋄ ".join(parts), grouping_mechanism if used_default else None)
 
-    return GroupingResult(mechanism.grouper(calculated_type, calculated_value, transaction), grouping_mechanism)
+    return GroupingResult(mechanism.grouper(data, calculated_type, calculated_value, transaction), grouping_mechanism)
 
 
 def get_issue_grouper_for_data(data, calculated_type=None, calculated_value=None, grouping_mechanism=None):
     return get_grouping_result_for_data(
         data, calculated_type, calculated_value, grouping_mechanism).grouping_key
-
-
-# TODO(#440): wire this in as a selectable grouping mechanism, not as an in-place default change.
-def get_issue_grouper_for_data_with_normalized_exception_value(data, calculated_type=None, calculated_value=None):
-    from issues.grouping_mechanisms.up_until_2_4_0 import default_issue_grouper
-
-    if calculated_type is None and calculated_value is None:
-        # convenience for calling code from tests, when digesting we don't do this because we already have this info
-        calculated_type, calculated_value = get_type_and_value_for_data(data)
-
-    transaction = force_str(data.get("transaction") or "<no transaction>")
-    fingerprint = data.get("fingerprint")
-
-    if "exception" in data and data["exception"]:
-        calculated_value = normalize_message_for_grouping(force_str(calculated_value))
-
-    if fingerprint:
-        return " ⋄ ".join([
-            (default_issue_grouper(calculated_type, calculated_value, transaction)
-             if part == "{{ default }}"
-             else str(part))
-            for part in fingerprint
-        ])
-
-    return default_issue_grouper(calculated_type, calculated_value, transaction)
 
 
 def get_title_for_exception_type_and_value(type_, value):
